# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COUNTLAH Invoice Processing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  message: string;
}

export interface MeResponse {
  authenticated: boolean;
  email?: string;
  isAdmin?: boolean;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LineItem {
  description: string;
  quantity: number;
  unit_price: number;
  amount: number;
}

export interface ExtractedData {
  supplier_name?: string;
  invoice_number?: string;
  invoice_date?: string;
  due_date?: string;
  currency?: string;
  subtotal?: number;
  tax?: number;
  total?: number;
  line_items?: LineItem[];
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  uploaded: "uploaded",
  extracted: "extracted",
  pushed: "pushed",
  failed: "failed",
} as const;

export interface Invoice {
  id: number;
  filename: string;
  originalName?: string;
  rawText?: string;
  extractedData?: ExtractedData;
  status: InvoiceStatus;
  createdAt: string;
  updatedAt?: string;
}

export interface InvoiceListResponse {
  invoices: Invoice[];
  total: number;
  page: number;
  limit: number;
}

export interface InvoiceStats {
  total: number;
  uploaded: number;
  extracted: number;
  pushed: number;
  failed: number;
  totalValue: number;
}

export type UpdateInvoiceBodyStatus =
  (typeof UpdateInvoiceBodyStatus)[keyof typeof UpdateInvoiceBodyStatus];

export const UpdateInvoiceBodyStatus = {
  uploaded: "uploaded",
  extracted: "extracted",
  pushed: "pushed",
  failed: "failed",
} as const;

export interface UpdateInvoiceBody {
  extractedData?: ExtractedData;
  status?: UpdateInvoiceBodyStatus;
}

export interface XeroPushResponse {
  success: boolean;
  xeroInvoiceId?: string;
  message: string;
}

export interface InvoiceLog {
  id: number;
  invoiceId: number;
  action: string;
  status: string;
  message?: string;
  createdAt: string;
}

export type ListInvoicesParams = {
  status?: ListInvoicesStatus;
  page?: number;
  limit?: number;
};

export type ListInvoicesStatus =
  (typeof ListInvoicesStatus)[keyof typeof ListInvoicesStatus];

export const ListInvoicesStatus = {
  uploaded: "uploaded",
  extracted: "extracted",
  pushed: "pushed",
  failed: "failed",
} as const;