# packages/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COUNTLAH Invoice Processing API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedData } from "./extractedData";
import type { InvoiceStatus } from "./invoiceStatus";

export interface Invoice {
  id: number;
  filename: string;
  originalName?: string;
  rawText?: string;
  extractedData?: ExtractedData;
  status: InvoiceStatus;
  createdAt: string;
  updatedAt?: string;
}