# packages/api-zod/src/generated/types/invoiceListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COUNTLAH Invoice Processing API
 * OpenAPI spec version: 0.1.0
 */
import type { Invoice } from "./invoice";

export interface InvoiceListResponse {
  invoices: Invoice[];
  total: number;
  page: number;
  limit: number;
}