# packages/api-zod/src/generated/types/lineItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COUNTLAH Invoice Processing API
 * OpenAPI spec version: 0.1.0
 */

export interface LineItem {
  description: string;
  quantity: number;
  unit_price: number;
  amount: number;
}