# packages/api-zod/src/generated/types/updateInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COUNTLAH Invoice Processing API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedData } from "./extractedData";
import type { UpdateInvoiceBodyStatus } from "./updateInvoiceBodyStatus";

export interface UpdateInvoiceBody {
  extractedData?: ExtractedData;
  status?: UpdateInvoiceBodyStatus;
}